[#5] refactor: types 폴더 분리
