+++ src/types/product.ts
@@ -1,0 +1,6 @@
+export interface Product {
+  title: string;
+  image: string;
+  price: string | number;
+  productId?: string;
+}
